Celestia persistent storage with PersistentVolumeClaim and StorageClass (#66)
Implements persistent storage for Celestia using the
PersistentVolumeClaim and StorageClass k8s resources.

Since we are using `kind` for local development, a local persistent
volume requires you to add an `extraMounts` section in your cluster
config to mount a host directory to the node. Now you can use this
directory in your `PersistentVolume`'s `local.path` setting.

On GC, we will need to define a `StorageClass`. The
`PersistentVolumeClaim` defines a storageclass, and the storageclass'
`provisioner` is responsible for dynamically provisioning the volume if
it does not exist.

Resolves https://github.com/astriaorg/dev-cluster/issues/27